add a new tag to a resorce
